requirements.txt
flask
fastapi
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
moviepy
ffmpeg-python
gtts
python-telegram-bot
twilio